Restructuring so my project is more readable.
