+++ src/common/constants/categories.js
@@ -1,0 +1,67 @@
+export const categories = [
+  "JAVA",
+  "PYTHON",
+  "JAVASCRIPT",
+  "C",
+  "C_PLUS_PLUS",
+  "C_SHARP",
+  "PHP",
+  "RUBY",
+  "SQL",
+  "SWIFT",
+  "GO",
+  "TYPESCRIPT",
+  "KOTLIN",
+  "RUST",
+  "SCALA",
+  "PERL",
+  "R",
+  "MATLAB",
+  "OBJECTIVE_C",
+  "VISUAL_BASIC",
+  "HASKELL",
+  "CLOJURE",
+  "GROOVY",
+  "DART",
+  "BASH",
+  "COBOL",
+  "FORTRAN",
+  "LISP",
+  "PASCAL",
+  "PROLOG",
+  "SCHEME",
+  "SMALLTALK",
+  "ADA",
+  "FORTH",
+  "F_SHARP",
+  "LUA",
+  "OCAML",
+  "PEARL",
+  "POWERSHELL",
+  "RACKET",
+  "VB_SCRIPT",
+  "VERILOG",
+  "VHDL",
+  "COFFEE_SCRIPT",
+  "ERLANG",
+  "HACK",
+  "JULIA",
+  "OBJECTIVE_J",
+  "PUPPET",
+  "RUBY_ON_RAILS",
+  "DJANGO",
+  "SPRING",
+  "FLASK",
+  "LARAVEL",
+  "EXPRESS_JS",
+  "ASP_NET",
+  "ANGULAR",
+  "REACT",
+  "VUE_JS",
+  "EMBER_JS",
+  "METEOR_JS",
+  "BACKBONE_JS",
+  "NODE_JS",
+  "RUBY_GEMS",
+  "DJANGO_REST_FRAMEWORK",
+].sort();
